+++ edge/pullRequest.ts
@@ -32,8 +32,9 @@ export default async function () {
 		throw `The current branch was out-of-sync with its remote branch.`
 	}
 	if (status.remote === '' || status.sync === Git.SyncStatus.LocalIsAheadOfRemote) {
-		const error = await push()
-		if (error !== undefined) {
+		const updated = await push()
+		if (updated === null) {
+			// Stop processing if the push operation is aborted
 			return null
 		}
 	}